bugfix: 🐛 jpnデータの修正  (37)
