+++ src/i18n/artifact/piece/jpn.json
