+++ src/i18n/artifact/set/jpn.json
