+++ src/i18n/artifact/subStatuses/jpn.json
@@ -2,9 +2,6 @@
   "hp": "HP",
   "atk": "攻撃力",
   "def": "防御力",
-  "hpPercent": "HP%",
-  "atkPercent": "攻撃力%",
-  "defPercent": "防御力%",
   "elementalMastery": "元素熟知",
   "energyRecharge": "元素チャージ効率",
   "critRate": "会心率",
