--- src/i18n/artifacts/piece/index.d.ts
@@ -1,9 +1,0 @@
-import type { ArtifactsPiece } from "../../../core/domain/artifacts.domain";
-
-type Piece = {
-  [key in ArtifactsPiece]: string;
-};
-
-declare const piece: Piece;
-
-export default piece;
