--- src/i18n/artifacts/set/index.d.ts
@@ -1,9 +1,0 @@
-import type { ArtifactsSet } from "../../../core/domain/artifacts.domain";
-
-type Set = {
-  [key in ArtifactsSet]: string;
-};
-
-declare const set: Set;
-
-export default set;
