--- src/i18n/artifacts/subStatus/index.d.ts
@@ -1,16 +1,0 @@
-interface SubStatus {
-  hp: string;
-  atk: string;
-  def: string;
-  hpPersent: string;
-  atkPersent: string;
-  defPersent: string;
-  elementalMastery: string;
-  energyRecharge: string;
-  critRate: string;
-  critDmg: string;
-}
-
-declare const subStatus: SubStatus;
-
-export default subStatus;
